backend/requirements.txt
Flask
pyopenssl
algoliasearch
python-dotenv
lark